Refactors after merging develop branch.
+ Renamed pipeline-lib package to be consistent with other libraries
+ Synched IDS Factory names in hazelcast configuration
