--- proven-member/pipeline-lib/src/main/java/gov/pnnl/cluster/lib/pipeline/gridappsd/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author d3a303
- *
- */
-package gov.pnnl.cluster.lib.pipeline.gridappsd;
