--- proven-member/pipeline-lib/src/main/java/gov/pnnl/cluster/lib/pipeline/service/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author d3j766
- *
- */
-package gov.pnnl.cluster.lib.pipeline.service;
