+++ proven-member/pipeline-lib/src/main/java/gov/pnnl/proven/cluster/lib/pipeline/PipelineIDSFactory.java
@@ -1,11 +1,11 @@
-package gov.pnnl.cluster.lib.pipeline;
+package gov.pnnl.proven.cluster.lib.pipeline;
 
 import com.hazelcast.nio.serialization.DataSerializableFactory;
 import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
 
-import gov.pnnl.cluster.lib.pipeline.response.MessageContentRoutingResponse;
-import gov.pnnl.cluster.lib.pipeline.response.TsResponse;
 import gov.pnnl.proven.cluster.lib.member.IDSFactory;
+import gov.pnnl.proven.cluster.lib.pipeline.response.MessageContentRoutingResponse;
+import gov.pnnl.proven.cluster.lib.pipeline.response.TsResponse;
 
 /**
  *
