+++ proven-member/pipeline-lib/src/main/java/gov/pnnl/proven/cluster/lib/pipeline/response/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author d3j766
+ *
+ */
+package gov.pnnl.proven.cluster.lib.pipeline.response;
